@@ -44,33 +44,35 @@ public class SecurityConf extends WebSecurityConfigurerAdapter {
     @Override
     protected void configure(HttpSecurity httpSecurity) throws Exception {
         httpSecurity
-                .csrf().disable()
                 .sessionManagement()
                 .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 .and()
+                .csrf().disable()
                 .authorizeRequests()
+
                 .antMatchers(HttpMethod.GET,"/").permitAll()
-                .antMatchers(HttpMethod.POST, "api/v1/user/auto").permitAll()
-                .antMatchers(HttpMethod.POST, "api/v1/user/register").permitAll()
-                .antMatchers(HttpMethod.GET,"api/v1/user/{id}").permitAll()
-                .antMatchers(HttpMethod.GET,"api/v1/user").permitAll()
+                .antMatchers(HttpMethod.POST, "/api/v1/user/auto").permitAll()
+                .antMatchers(HttpMethod.POST, "/api/v1/user/register").permitAll()
+                .antMatchers(HttpMethod.GET,"/api/v1/user/{id}").permitAll()
+                .antMatchers(HttpMethod.GET,"/api/v1/user").permitAll()
 
-                .antMatchers(HttpMethod.GET,"api/common-reference/get-cities").permitAll()
-                .antMatchers(HttpMethod.GET,"api/common-reference/get-areas").permitAll()
+                .antMatchers(HttpMethod.GET,"/api/common-reference/get-cities").permitAll()
+                .antMatchers(HttpMethod.GET,"/api/common-reference/get-areas").permitAll()
 
-                .antMatchers(HttpMethod.POST,"api/complex/save").hasAnyRole(SecurityRole.ROLE_USER .getName(),SecurityRole.ROLE_PROVIDER.getName())
-                .antMatchers(HttpMethod.GET,"api/complex/{id}").permitAll()
-                .antMatchers(HttpMethod.GET,"api/complex").permitAll()
-                .antMatchers(HttpMethod.GET,"api/complex/get-by-user-id/{id}").hasRole(SecurityRole.ROLE_PROVIDER.getName())
+                .antMatchers(HttpMethod.GET,"/api/complex/{id}").permitAll()
+                .antMatchers(HttpMethod.GET,"/api/complex").permitAll()
+                .antMatchers(HttpMethod.GET,"/api/complex/get-by-user-id/{id}").hasRole(SecurityRole.ROLE_PROVIDER.getName())
+                .antMatchers(HttpMethod.POST,"/api/complex/save").hasAnyRole(SecurityRole.ROLE_USER .getName(),SecurityRole.ROLE_PROVIDER.getName())
 
-                .antMatchers(HttpMethod.POST,"api/file/save").hasRole(SecurityRole.ROLE_PROVIDER.getName())
-                .antMatchers(HttpMethod.GET,"api/file/get-by-complex-id").permitAll()
+                .antMatchers(HttpMethod.GET,"/api/file/get-by-complex-id").permitAll()
+                .antMatchers(HttpMethod.POST,"/api/file/save").hasRole(SecurityRole.ROLE_PROVIDER.getName())
 
-                .antMatchers(HttpMethod.POST,"api/review/save").hasRole(SecurityRole.ROLE_USER.getName())
-                .antMatchers(HttpMethod.GET,"api/review/get-by-complex-id/{id}").permitAll()
-                .antMatchers(HttpMethod.GET,"api//review/get-all-by-complex-id/{id}").permitAll()
+                .antMatchers(HttpMethod.GET,"/api/review/get-by-complex-id/{id}").permitAll()
+                .antMatchers(HttpMethod.GET,"/api/review/get-all-by-complex-id/{id}").permitAll()
+                .antMatchers(HttpMethod.POST,"/api/review/save").hasRole(SecurityRole.ROLE_USER.getName())
 
                 .antMatchers(HttpMethod.GET,"/swagger-ui.html").permitAll()
+
                 .and()
                 .httpBasic();
     }
