@@ -1,14 +1,18 @@
 package com.example.yssyk_cool.controller;
 
+import com.example.yssyk_cool.dto.complex.request.ComplexForUpdateRequest;
 import com.example.yssyk_cool.dto.complex.request.ComplexRequest;
 import com.example.yssyk_cool.dto.complex.response.ComplexResponse;
 import com.example.yssyk_cool.service.ComplexService;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.FileNotFoundException;
 import java.util.List;
 
 @RestController
@@ -22,28 +26,44 @@ public class ComplexController {
     final ComplexService complexService;
 
     @PostMapping("/save")
-    public ComplexResponse save(@RequestBody ComplexRequest complexRequest) {
+    @ApiOperation("Сохранение комплекса")
+    public ComplexResponse save(@RequestBody ComplexRequest complexRequest) throws FileNotFoundException {
         return complexService.save(complexRequest);
     }
 
     @GetMapping("/{id}")
-    public ComplexResponse getById(@PathVariable Long id) {
+    @ApiOperation("Получить комплес по id")
+    public ComplexResponse getById(@PathVariable Long id) throws FileNotFoundException {
         return complexService.findById(id);
     }
-
     @GetMapping
+    @ApiOperation("Получить все комплексы")
     public List<ComplexResponse> getAll() {
         return complexService.getAll();
     }
 
     @GetMapping("/get-by-user-id/{id}")
+    @ApiOperation("Получить все посты одношо провайдера по его id")
     public List<ComplexResponse> getByUserId(@PathVariable Long id){
         return complexService.findAllByUserId(id);
     }
 
     @GetMapping("/search/{search}")
+    @ApiOperation("Поиск")
     public List<ComplexResponse> search(@PathVariable String search) {
-        return null;
+        return complexService.search(search);
+    }
+
+    @DeleteMapping("/delete/{id}")
+    @ApiOperation("Удалить комплек по id")
+    public ComplexResponse delete(@PathVariable("id")Long id){
+        return complexService.delete(id);
+    }
+
+    @PutMapping("/update")
+    @ApiOperation("Обновить комплекс")
+    public ComplexResponse update(@RequestBody ComplexForUpdateRequest complexRequest){
+        return complexService.update(complexRequest);
     }
 }
 
