@@ -4,12 +4,15 @@ import com.example.yssyk_cool.dto.review.request.ReviewRequest;
 import com.example.yssyk_cool.dto.review.response.ReviewResponse;
 import com.example.yssyk_cool.entity.Complex;
 import com.example.yssyk_cool.service.ReviewService;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.FileNotFoundException;
 import java.util.List;
 
 @RestController
@@ -23,16 +26,13 @@ public class ReviewController {
     final ReviewService reviewService;
 
     @PostMapping("/save")
-    public ReviewResponse save(@RequestBody ReviewRequest request){
+    @ApiOperation("Добавить отзыв для комплекса")
+    public ReviewResponse save(@RequestBody ReviewRequest request) throws FileNotFoundException {
         return reviewService.save(request);
     }
 
-    @GetMapping("/{id}")
-    public ReviewResponse getByComplexId(@PathVariable("id") Long id){
-        return reviewService.findById(id);
-    }
-
     @GetMapping("/get-all-by-complex-id/{id}")
+    @ApiOperation("Получить все отзывы по id комплекса")
     public List<ReviewResponse> getAllByComplexId(@PathVariable("id") Complex complex){
         return reviewService.getAllByComplexId(complex);
     }
