@@ -6,6 +6,8 @@ import com.example.yssyk_cool.dto.user.response.UserResponse;
 import com.example.yssyk_cool.dto.user.response.UserTokenResponse;
 import com.example.yssyk_cool.exception.EmailNotBeEmptyException;
 import com.example.yssyk_cool.service.auth.UserService;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 
@@ -13,6 +15,7 @@ import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.FileNotFoundException;
 import java.util.List;
 
 @RestController
@@ -27,11 +30,13 @@ public class UserController {
 
 
     @PostMapping("/register")
-    public UserResponse register(@RequestBody UserRequest request) throws EmailNotBeEmptyException {
+    @ApiOperation("Регистрация")
+    public UserResponse register(@RequestBody UserRequest request) throws EmailNotBeEmptyException, FileNotFoundException {
         return userService.save(request);
     }
 
     @PostMapping("/auto")
+    @ApiOperation("Авторизация")
     public UserTokenResponse auto(@RequestBody UserAuthRequest request) {
         return userService.getToken(request);
     }
@@ -43,7 +48,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public UserResponse getUserById(@PathVariable("id") Long id) {
+    public UserResponse getUserById(@PathVariable("id") Long id) throws FileNotFoundException {
         return userService.findById(id);
     }
 
