@@ -1,6 +1,8 @@
 package com.example.yssyk_cool.dto.complex.request;
 
+import com.example.yssyk_cool.dto.contactInfo.request.ContactInfoForUpdateRequest;
 import com.example.yssyk_cool.dto.contactInfo.request.ContactInfoRequest;
+import com.example.yssyk_cool.dto.location.request.LocationForUpdateRequest;
 import com.example.yssyk_cool.dto.location.request.LocationRequest;
 import com.example.yssyk_cool.enums.TypeComplex;
 import lombok.*;
@@ -13,9 +15,10 @@ import javax.validation.constraints.NotNull;
 @NoArgsConstructor
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class ComplexRequest {
+public class ComplexForUpdateRequest {
 
-    Long createdById;
+    Long complexId;
+
 
     @NotNull
     String nameComplex;
@@ -26,8 +29,7 @@ public class ComplexRequest {
 
     TypeComplex typeComplex;
 
-    LocationRequest locationRequest;
+    LocationForUpdateRequest locationRequest;
 
-    ContactInfoRequest contactInfoRequest;
-
+    ContactInfoForUpdateRequest contactInfoRequest;
 }
